+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This version add the orders module
 - Added endpoints for CRUD orders and list templates in `src/api/v1/apps/orders/router.py`
 - Added endpoint for products list in `src/api/v1/apps/orders/router.py`
 - Added endpoint for flavors list in `src/api/v1/apps/orders/router.py`
+- Added delete items from orders endpoint in `src/api/v1/apps/orders/router.py`
 
 ### Changed
 - Added current_user at context in patch.products endpoint in `src/api/v1/apps/orders/router.py`